step completion improvements
get better completion suggestions for steps, and also insert only the part of a step that has not already been typed.

implementation of Biometria-se/grizzly#166, support for definition, and if step expressions contains a reference to a request payload template file, it is now possible to ctrl + click it to open it up.