completion improvements
fixed problem in Biometria-se/grizzly#296.

implemented better support for `And` and `But`, since they depend on a non-"one of those" steps arbritary number of lines above. now when completing with those, it will only return suggestions related to the identified "base" keyword above it, instead of all.

client/vscode completion tests has been greatly approved by actually accepting a suggestion and verifying that the complete line turned out as expected. that is a hack though, since there are no events to listen for that would tell you that the UI is done, so sleeps are introduced :(
one of the sleeps could be removed by adding a onDidTextDocumentChange listener, but it should be removed after each time it has executed, and then it becomes a chicken-and-egg problem...