fixed problems with selecting correct python environment
removed debug logging.

ms-python.python extension, which provide the @vscode/python-extension API had problems selecting the correct python environment, causing the tests to fail since the language server couldn't be started. the language server module was installed in the environment action/setup-python had configured.
in the runner a version of 3.8 through 3.12 was installed, and for some reason the extension (on windows and mac) decided that the active environment was the highest version...