completion improvements (#51)
fixed problem in Biometria-se/grizzly#296.

implemented better support for `And` and `But`, since they depend on a non-"one of those" steps arbritary number of lines above. now when completing with those, it will only return suggestions related to the identified "base" keyword above it, instead of all.

`client/vscode` completion tests has been greatly approved by actually accepting a suggestion and verifying that the completef line turned out as expected. that is a hack though, since there are no events to listen for that would tell you that the UI is done, so sleeps are introduced :(

`ms-python.python` and it's `@vscode/python-extension` API is buggy when it comes to selecting python environment, seems like it does not take which python environment that is active when executing into consideration at all.

workaround this in the extension by looking if `VIRTUAL_ENV` environment variable is set, and use the value of that to resolve the environment used.

in the workflow `action/setup-python` will set environment variable `pythonLocation` for the environment that is configured. so when running the client test we set `VIRTUAL_ENV` to the value of `pythonLocation`.